requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Migrate
python-dotenv
pandas
openpyxl
qrcode[pil]
Pillow
ReportLab
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
gunicorn
pytest
ruff
black
